+++ pkg/innerring/processors/netmap/nodevalidation/locode/validator.go
@@ -17,23 +17,13 @@ type Validator struct {
 
 // New creates a new instance of the Validator.
 //
-// Panics if at least one value of the parameters is invalid.
-//
 // The created Validator does not require additional
 // initialization and is completely ready for work.
 func New() *Validator {
 	return &Validator{}
 }
 
-func (v *Validator) Get(lc string) (*locodedb.Key, locodedb.Record, error) {
-	country, location := lc[:2], lc[2:]
-	if lc[2] == ' ' {
-		location = lc[3:]
-	}
-	key, err := locodedb.NewKey(country, location)
-	if err != nil {
-		return nil, locodedb.Record{}, err
-	}
+func getRecord(lc string) (locodedb.Record, error) {
 	rec, err := locodedb.Get(lc)
-	return key, rec, err
+	return rec, err
 }