+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog for NeoFS Node
 - New optimized SDK version is integrated (#2622)
 - IR uses internal LOCODE DB from Go package (#2610)
 - Contract group for the committee is no longer registered/used in the Sidechain auto-deployment (#2642)
+- IR does not change SNs' attributes, SNs expand UN/LOCODE attributes do it themselves (#2612)
 - The priority of metrics service is increased (#2428)
 - The priority of running control service is maximized (#2585)